fix: call render function from backend
