+++ frontend/express/public/javascripts/countly/countly.helpers.js
@@ -337,7 +337,12 @@
         }
         var payload = {};
         var persistent = msg.persistent;
-        payload.text = countlyCommon.encodeHtml(msg.message);
+        if (msg.html) {
+            payload.text = msg.message;
+        }
+        else {
+            payload.text = countlyCommon.encodeHtml(msg.message);
+        }
         payload.autoHide = !msg.sticky;
         payload.id = msg.id;
         payload.width = msg.width;
